Move wasm to a top-level package and de-duplicate
