+++ .gitignore
@@ -67,7 +67,7 @@ apps/site/test-results
 packages/block-template-html/dev/src
 packages/mock-block-dock/src/version.ts
 packages/@blockprotocol/type-system/.rollup.cache
-packages/@blockprotocol/type-system/src/wasm
+packages/@blockprotocol/type-system/wasm
 
 ##################
 ## Specific to git
