+++ .prettierignore
@@ -88,4 +88,4 @@ apps/site/site-map.json
 apps/site/test-results
 packages/block-template-html/dev/src
 packages/mock-block-dock/src/version.ts
-packages/@blockprotocol/type-system/src/wasm
+packages/@blockprotocol/type-system/wasm
