+++ packages/@blockprotocol/type-system/src/index.ts
@@ -1,8 +1,8 @@
+import wasm from "../wasm/type-system_bg.wasm";
 import { setWasmInit } from "./common";
-import wasm from "./wasm/type-system_bg.wasm";
 
+export * from "../wasm/type-system";
 export { TypeSystemInitializer } from "./common";
-export * from "./wasm/type-system";
 
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-expect-error
